+++ src/operator/tensor/index_add_forward.cu
@@ -38,7 +38,7 @@ struct IndexAddForwardGPUKernel {
                                   const mshadow::Shape<MXNET_SPECIAL_MAX_NDIM>& a_pre_stride,
                                   const mshadow::Shape<MXNET_SPECIAL_MAX_NDIM>& val_stride,
                                   const mshadow::Shape<MXNET_SPECIAL_MAX_NDIM>& val_shape,
-                                  const size_t a_tail_size, const int ind_num,
+                                  const int a_tail_size, const int ind_num,
                                   const int ind_ndim, const int* ind,
                                   const int a_ndim) {
     size_t id = 0;
@@ -47,7 +47,7 @@ struct IndexAddForwardGPUKernel {
       id += a_pre_stride[seg + dim] * ind[dim * ind_num + i];
     }
     id *= a_tail_size;
-    for (size_t _i = 0; _i < a_tail_size; ++_i) {
+    for (int _i = 0; _i < a_tail_size; ++_i) {
       mshadow::Shape<MXNET_SPECIAL_MAX_NDIM> a_tail_id = mxnet_op::unravel(_i, a_tail_shape);
       mshadow::Shape<MXNET_SPECIAL_MAX_NDIM> val_id;
       for (int _j = seg; _j < seg + a_ndim; ++_j) {
@@ -69,7 +69,7 @@ void IndexAddForwardCalc(mshadow::Stream<xpu> *s,
                          const mshadow::Shape<MXNET_SPECIAL_MAX_NDIM>& val_stride,
                          const mshadow::Shape<MXNET_SPECIAL_MAX_NDIM>& val_shape,
                          const mshadow::Shape<MXNET_SPECIAL_MAX_NDIM>& a_shape,
-                         const size_t a_tail_size,
+                         const int a_tail_size,
                          const int ind_ndim, const int* ind,
                          const int a_ndim) {
   using namespace mxnet_op;
